requirements/base.txt
Django==5.1.1
djangorestframework==3.15.2
python-environ==0.4.54
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
django-debug-toolbar==4.4.6
django-cors-headers==4.4.0